+++ app/services/analysis/gwas/genotype_hapmap_metadata_analyzer.rb
@@ -1,0 +1,22 @@
+require "csv"
+
+class Analysis
+  module Gwas
+    class GenotypeHapmapMetadataAnalyzer
+      def call(path)
+        File.open(path, "r") do |hapmap_file|
+          hapmap_data = Analysis::Gwas::GenotypeHapmapParser.new.call(hapmap_file)
+          chromosomes = Set.new
+
+          raise unless hapmap_data.valid?
+
+          hapmap_data.each_record do |record|
+            chromosomes << record.chrom
+          end
+
+          [hapmap_data.sample_ids, chromosomes.to_a]
+        end
+      end
+    end
+  end
+end
