+++ app/services/analysis/gwas/genotype_hapmap_to_csv_converter.rb
@@ -1,11 +1,11 @@
 require "csv"
 
 class Analysis
-  class Gwasser
+  module Gwas
     class GenotypeHapmapToCsvConverter
       def call(path)
         File.open(path, "r") do |hapmap_file|
-          hapmap_data = Analysis::GenotypeHapmapParser.new.call(hapmap_file)
+          hapmap_data = Analysis::Gwas::GenotypeHapmapParser.new.call(hapmap_file)
 
           raise unless hapmap_data.valid?
