+++ app/services/analysis/gwas/genotype_vcf_to_csv_converter.rb
@@ -1,11 +1,11 @@
 require "csv"
 
 class Analysis
-  class Gwasser
+  module Gwas
     class GenotypeVcfToCsvConverter
       def call(path)
         File.open(path, "r") do |vcf_file|
-          vcf_data = Analysis::GenotypeVcfParser.new.call(vcf_file)
+          vcf_data = Analysis::Gwas::GenotypeVcfParser.new.call(vcf_file)
 
           raise unless vcf_data.valid?
