+++ app/services/analysis/gwas/setup/phenotype_plant_trial.rb
@@ -14,7 +14,7 @@ class Analysis
 
         def call
           plant_trial = PlantTrial.visible(analysis.owner_id).find(analysis.meta.fetch("plant_trial_id"))
-          pheno_csv_file, trait_names = Analysis::PlantTrialPhenotypeCsvBuilder.new.build_csv(plant_trial)
+          pheno_csv_file, trait_names = Analysis::Gwas::PlantTrialPhenotypeCsvBuilder.new.build_csv(plant_trial)
 
           traits_to_remove, _ = analyze_pheno_csv_file(pheno_csv_file)
