+++ spec/services/analysis/gwas/phenotype_csv_parser_spec.rb
@@ -1,19 +1,19 @@
 require "rails_helper"
 
-RSpec.describe Analysis::GenotypeCsvParser do
+RSpec.describe Analysis::Gwas::PhenotypeCsvParser do
 
   subject { described_class.new.call(input) }
 
   context "with valid input" do
-    let(:input) { fixture_file("gwas-genotypes.csv", "text/csv") }
-    let(:headers) { %w(ID) + 1.upto(500).map { |id| "snp-#{id}" } }
+    let(:input) { fixture_file("gwas-phenotypes.csv", "text/csv") }
+    let(:headers) { %w(ID) + 1.upto(10).map { |id| "trait #{id}" } }
 
     it "returns valid result object" do
       expect(subject).to be_valid
       expect(subject.headers).to eq(headers)
       expect(subject.csv.pos).to eq(subject.rewind(skip_header: true))
-      expect(subject.csv.readline[0]).to eq("plant-1")
-      expect(subject.csv.readline.size).to eq(501)
+      expect(subject.csv.readline[0]).to eq("plant-6")
+      expect(subject.csv.readline.size).to eq(1 + 10)
     end
   end
 
@@ -23,7 +23,7 @@ RSpec.describe Analysis::GenotypeCsvParser do
     it "returns invalid result object" do
       expect(subject).not_to be_valid
       expect(subject.errors).to include(:no_id_column)
-      expect(subject.errors).to include(:no_mutations)
+      expect(subject.errors).to include(:no_traits)
       expect(subject.errors).to include(:no_samples)
     end
   end
