+++ spec/services/analysis/gwas/plant_trial_phenotype_csv_builder_spec.rb
@@ -1,6 +1,6 @@
 require "rails_helper"
 
-RSpec.describe Analysis::PlantTrialPhenotypeCsvBuilder do
+RSpec.describe Analysis::Gwas::PlantTrialPhenotypeCsvBuilder do
   let(:plant_trial) { create(:plant_trial) }
   let!(:plant_scoring_units) { create_list(:plant_scoring_unit, 3, plant_trial: plant_trial) }
   let!(:trait_scores) {
